STCLI-163 depend on nightmare-less stripes-testing
`nightmare` was removed from `@folio/stripes-testing` in `v3`. We should
yank the related hooks out here too so we don't have a bunch
of do-nothing commands, but actually removing nightmare, along with the
ancient dependencies (electron 2!!!) it drags along, was an even higher
priority, even if it leaves with some (unused) do-nothing commands.

Refs STCLI-163