STCLI-188 avoid sabotaged colors
Avoid the sabotaged `colors` module by adding a dev-dep on the
last-known-good build, `1.4.0`. Using a dev-dep, even though `colors` is
technically only a transitive dep, allows both `yarn` and `npm` to
handle this change. In yarn-land, arguably a `resolutions` entry would
be more correct, but that would only work in yarn-land.

Refs STCLI-188