STCLI-193 avoid sabotaged isBinaryFile
A breaking change to node compatibilty for `isBinaryFile`, a transitive
dependency via `karma`, was published in a patch release. It isn't clear
if the author doesn't understand semantic versioning or just doesn't
care about it, but multiple pleas to move the changes to a new major
version and publish a patch to restore compatibility in the 4.0.x line
were rebuffed (https://github.com/gjtorikian/isBinaryFile/issues/51).

Refs STCLI-193