STCLI-195 avoid webpack > 5.69.1
`webpack` `5.69.0` and `5.70.1` introduced changes to "handle multiple
alternative directories (e. g. due to resolve.alias or resolve.modules)
when creating an context module" that appear to be incompatible with the
way moment handles dynamic loading of locales. That change was reverted
in `5.69.1` and then re-applied in `5.70.0`, hence the choice of
`5.69.1` here.

Honestly, the details are fuzzy but the incompatibility (in the form of
~7k lines of warnings) is indisputable. I believe the issue is captured
by https://github.com/moment/moment/issues/5492. It feels like `moment`
was built in a pre-ESM world and just can't cope with how things have
evolved since.

Refs STCLI-195