Bump sinon to 15.0.4, sinon-chai to 3.7.0 (#324)
Bump sinon from 4.1.4 to 15.0.4. Breaking changes in sinon are not well-documented, even though CHANGES.md catalogs many new major versions. Given this is only a dev-dep and our tests still run with just a minor tweak, this implies that whatever surfaces did change between versions 4 and 15 are not surfaces that we touched. Whew.

Bump its companion sinon-chai from 2.14.0 to 3.7.0. The only breaking change here was in raising the minimum versions of node, sinon, and chai to versions we already require.