STCLI-235 update stripes-webpack to v5
*BREAKING* Bump `@folio/stripes-webpack` to v5. stripes-cli re-exports
portions of stripes-webpack and, from the point of view of applications
with dev-deps on stripes-cli, represents the public API for
stripes-webpack. Thus, even though stripes-webpack is a direct (and thus
internal) dependency here, it functions more like a peer-dependency and
since that version is changing, then this one must too.

Refs STCLI-235