Revert "STCLI-260 bump NodeJS to v20 (#371)" (#376)
This reverts commit 69b4aef7df3204731eb8dd464162b23a68147096.

The Jenkins nodes where we build platform-complete don't yet have NodeJS
v20 and we don't know when it will be available. In the interest of
making other changes (which, AFAIK, do directly not rely on NodeJS v20)
available ASAP, we are reverting this change until our build nodes can
support it.

Refs [STCLI-260](https://folio-org.atlassian.net/browse/STCLI-260)